% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_helper.R
\name{tb_helper}
\alias{tb_helper}
\title{Help for reading table builder files with 10 Year Groups and Sex}
\usage{
tb_helper()
}
\value{

}
\description{
Help for reading table builder files with 10 Year Groups and Sex
}
